requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
boto3==1.22.9
botocore==1.25.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
datasets==2.1.0
dill==0.3.4
docker-pycreds==0.4.0
filelock==3.6.0
fonttools==4.32.0
frozenlist==1.3.0
fsspec==2022.3.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.3
jmespath==1.0.0
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==6.0.2
multiprocess==0.70.12.2
numpy==1.18.5
packaging==21.3
pandas==1.0.5
path==16.2.0
pathtools==0.1.2
Pillow==9.0.1
pip==22.0.4
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyarrow==8.0.0
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
pytz==2021.3
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
responses==0.18.0
rouge==1.0.1
s3transfer==0.5.2
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sentencepiece==0.1.96
sentry-sdk==1.5.8
setproctitle==1.2.2
setuptools==58.0.4
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
SoundFile==0.10.3.post1
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.64.0
transformers==4.17.0
typing-extensions==3.10.0.2
urllib3==1.26.9
wandb==0.12.11
wheel==0.37.1
wincertstore==0.2
xxhash==3.0.0
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
